resolved all issues with postgresql interface
