+++ hmda-regulator/src/main/resources/application.conf
@@ -18,7 +18,7 @@ akka {
 hmda {
   regulator {
     http {
-      createSchema = false
+      createSchema = true
       createSchema = ${?INSTITUTION_PG_CREATE_SCHEMA}
       timeout = 3
       host = ${?HTTP_REGULATOR_HOST}
