--- hmda-regulator/src/main/scala/hmda/regulator/data/RegulatorDataEntity.scala
@@ -1,17 +1,0 @@
-package hmda.regulator.data
-
-class RegulatorDataEntity() {
-  private var _dataType: String = ""
-
-  def dataType: String = _dataType
-
-  def dataType_=(value: String) = {
-    _dataType = value
-  }
-
-  def toCSV: String = {
-
-    s"testing|1|32|3"
-  }
-
-}
