--- hmda-regulator/src/main/scala/hmda/regulator/data/model/LarRegulatorData.scala
@@ -1,5 +1,0 @@
-package hmda.regulator.data.model
-
-import hmda.regulator.data.RegulatorDataEntity
-
-class LarRegulatorData extends RegulatorDataEntity {}
