--- hmda-regulator/src/main/scala/hmda/regulator/data/model/TsRegulatorData.scala
@@ -1,5 +1,0 @@
-package hmda.regulator.data.model
-
-import hmda.regulator.data.RegulatorDataEntity
-
-class TsRegulatorData extends RegulatorDataEntity {}
