--- hmda-regulator/src/main/scala/hmda/regulator/query/InstitutionEmailEntity.scala
@@ -1,7 +1,0 @@
-package hmda.regulator.query
-
-case class InstitutionEmailEntity(
-    id: Int = 0,
-    lei: String,
-    emailDomain: String
-)
