+++ hmda-regulator/src/main/scala/hmda/regulator/query/RegulatorComponent.scala
@@ -1,0 +1,92 @@
+package hmda.regulator.query
+
+import hmda.query.DbConfiguration._
+import hmda.query.repository.TableRepository
+import slick.basic.DatabaseConfig
+import slick.jdbc.JdbcProfile
+
+import scala.concurrent.Future
+
+trait RegulatorComponent {
+
+  import dbConfig.profile.api._
+
+  class InstitutionsTable(tag: Tag)
+      extends Table[InstitutionEntity](tag, "institutions2018") {
+    def lei = column[String]("lei", O.PrimaryKey)
+    def activityYear = column[Int]("activity_year")
+    def agency = column[Int]("agency")
+    def institutionType = column[Int]("institution_type")
+    def id2017 = column[String]("id2017")
+    def taxId = column[String]("tax_id")
+    def rssd = column[Int]("rssd")
+    def respondentName = column[String]("respondent_name")
+    def respondentState = column[String]("respondent_state")
+    def respondentCity = column[String]("respondent_city")
+    def parentIdRssd = column[Int]("parent_id_rssd")
+    def parentName = column[String]("parent_name")
+    def assets = column[Int]("assets")
+    def otherLenderCode = column[Int]("other_lender_code")
+    def topHolderIdRssd = column[Int]("topholder_id_rssd")
+    def topHolderName = column[String]("topholder_name")
+    def hmdaFiler = column[Boolean]("hmda_filer")
+
+    override def * =
+      (lei,
+       activityYear,
+       agency,
+       institutionType,
+       id2017,
+       taxId,
+       rssd,
+       respondentName,
+       respondentState,
+       respondentCity,
+       parentIdRssd,
+       parentName,
+       assets,
+       otherLenderCode,
+       topHolderIdRssd,
+       topHolderName,
+       hmdaFiler) <> (InstitutionEntity.tupled, InstitutionEntity.unapply)
+  }
+  val institutionsTable = TableQuery[InstitutionsTable]
+
+  class InstitutionRepository(val config: DatabaseConfig[JdbcProfile])
+      extends TableRepository[InstitutionsTable, String] {
+
+    override val table: config.profile.api.TableQuery[InstitutionsTable] =
+      institutionsTable
+
+    override def getId(row: InstitutionsTable): config.profile.api.Rep[Id] =
+      row.lei
+
+    def createSchema() = db.run(table.schema.create)
+    def dropSchema() = db.run(table.schema.drop)
+
+    def insert(institution: InstitutionEntity): Future[Int] = {
+      db.run(table += institution)
+    }
+
+    def findByLei(lei: String): Future[Seq[InstitutionEntity]] = {
+      db.run(table.filter(_.lei === lei).result)
+    }
+
+    def findActiveFilers(): Future[Seq[InstitutionEntity]] = {
+      db.run(table.filter(_.hmdaFiler === true).result)
+    }
+
+    def getAllInstitutions(): Future[Seq[InstitutionEntity]] = {
+      db.run(table.result)
+    }
+
+    def deleteByLei(lei: String): Future[Int] = {
+      db.run(table.filter(_.lei === lei).delete)
+    }
+
+    def count(): Future[Int] = {
+      db.run(table.size.result)
+    }
+  }
+
+}
